conversion from ‘int’ to ‘uint16_t’ {aka ‘short unsigned int’} may change value [-Werror=conversion]

+++ example_14/hardware/include/ros2_control_demo_example_14/rrbot_sensor_for_position_feedback.hpp
@@ -83,7 +83,7 @@ private:
 
   // Fake "mechanical connection" between actuator and sensor using sockets
   struct sockaddr_in address_;
-  int socket_port_;
+  uint16_t socket_port_;
   int address_length_;
   int obj_socket_;
   int sockoptval_ = 1;
